src/agent_backend/requirements.txt
# Core runtime
semantic-kernel[mcp]
python-dotenv
fastapi
uvicorn
gunicorn
pydantic
azure-cosmos >= 4.4.0
azure-identity
